Move dict<->graph conversion to backend base
