+++ gundb/backends/rediskv.py
@@ -4,8 +4,6 @@ from .utils import defaultify, fix_lists, desolve
 from ..consts import METADATA, SOUL, STATE
 import json
 ignore = '_'
-LISTDATA = 'L'
-MAPPING  = '[]'
 
 def format_object_id(schema, id):
     return "{}://{}".format(schema, id)
@@ -32,39 +30,6 @@ class RedisKV(BackendMixin):
         obj = self.delegate_list_metadatata(obj)
         self.redis.set(full_id, json.dumps(obj))
 
-    def delegate_list_metadatata(self, obj):
-        if not isinstance(obj, dict):
-            return obj
-        for k, v in obj.items():
-            if k == METADATA:
-                continue
-            if k.startswith("list_"):
-                obj[k] = self.delegate_list_metadatata(v)
-                obj[METADATA][LISTDATA][k][METADATA] = v[METADATA]
-                mapping, result_list = self.extract_mapping_list(obj[k])
-                obj[METADATA][LISTDATA][k][MAPPING] = mapping
-                obj[k] = result_list
-            else:
-                obj[k] = self.delegate_list_metadatata(v)
-        return obj
-
-    def extract_mapping_list(self, list_obj):
-        mapping = {}
-        result = []
-        del list_obj[METADATA]
-        number_of_nones = 0
-        for i, k in enumerate(list_obj.keys()):
-            index = result.index(list_obj[k]) if list_obj[k] in result else -1
-            if list_obj[k] == None:
-                number_of_nones += 1
-                mapping[k] = -1
-            elif index != -1:
-                mapping[k] = index
-            else:
-                mapping[k] = i - number_of_nones
-                result.append(list_obj[k])
-        return mapping, result
-
     def recover_graph(self):
         root_souls = self.redis.keys(pattern="*://*")
         graph = {}
@@ -77,27 +42,6 @@ class RedisKV(BackendMixin):
         db_form = defaultify(json.loads(self.redis.get(key)))
         return self.convert_to_graph(db_form)
     
-    def convert_to_graph(self, obj):
-        if not isinstance(obj, dict):
-            return obj
-        result = defaultify({})
-        obj = self.eliminate_lists(obj)
-        for k, v in obj.items():
-            result[k] = self.convert_to_graph(v)
-        return result
-
-    def eliminate_lists(self, obj):
-        if METADATA not in obj or LISTDATA not in obj[METADATA]:
-            return obj
-        for k, v in obj[METADATA][LISTDATA].items():
-            recovered_list = {METADATA: v[METADATA]}
-            for orig_key, i in v[MAPPING].items():
-                if i != -1:
-                    recovered_list[orig_key] = obj[k][i]
-            obj[k] = recovered_list
-        del obj[METADATA][LISTDATA]
-        return obj
-
     def __setitem__(self, k, v):
         self.db[k] = v
